moved rayon utils to concurrency module (#7974)
It will be reused for edge signature verification.
