+++ chain/network/src/accounts_data/mod.rs
@@ -24,83 +24,18 @@
 //!       discarding the progress
 //!     - banning a peer wouldn't help since peers are anonymous, so a single attacker can act as a
 //!       lot of peers
+use crate::concurrency;
 use crate::concurrency::arc_mutex::ArcMutex;
 use crate::network_protocol;
 use crate::network_protocol::SignedAccountData;
 use crate::types::AccountKeys;
-use near_o11y::log_assert;
 use near_primitives::types::{AccountId, EpochId};
-use rayon::iter::{ParallelBridge, ParallelIterator};
+use rayon::iter::ParallelBridge;
 use std::collections::HashMap;
-use std::future::Future;
-use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 
 #[cfg(test)]
 mod tests;
-
-struct MustCompleteGuard;
-
-impl Drop for MustCompleteGuard {
-    fn drop(&mut self) {
-        log_assert!(false, "dropped a non-abortable future before completion");
-    }
-}
-
-/// must_complete wraps a future, so that it logs an error if it is dropped before completion.
-/// Possibility of future abort at every await makes the control flow unnecessarily complicated.
-/// In fact, only few basic futures (like io primitives) actually need to be abortable, so
-/// that they can be put together into a tokio::select block. All the higher level logic
-/// would greatly benefit (in terms of readability and bug-resistance) from being non-abortable.
-/// Rust doesn't support linear types as of now, so best we can do is a runtime check.
-/// TODO(gprusak): we would like to make the futures non-abortable, however with the current
-/// semantics of actix, which drops all the futures when stopped this is not feasible.
-/// Reconsider how to introduce must_complete to our codebase.
-#[allow(dead_code)]
-fn must_complete<Fut: Future>(fut: Fut) -> impl Future<Output = Fut::Output> {
-    let guard = MustCompleteGuard;
-    async move {
-        let res = fut.await;
-        let _ = std::mem::ManuallyDrop::new(guard);
-        res
-    }
-}
-
-/// spawns a closure on a global rayon threadpool and awaits its completion.
-/// Returns the closure result.
-/// WARNING: panicking within a rayon task seems to be causing a double panic,
-/// and hence the panic message is not visible when running "cargo test".
-async fn rayon_spawn<T: 'static + Send>(f: impl 'static + Send + FnOnce() -> T) -> T {
-    let (send, recv) = tokio::sync::oneshot::channel();
-    rayon::spawn(move || {
-        if send.send(f()).is_err() {
-            tracing::warn!("rayon_spawn has been aborted");
-        }
-    });
-    recv.await.unwrap()
-}
-
-/// Applies f to the iterated elements and collects the results, until the first None is returned.
-/// Returns the results collected so far and a bool (false iff any None was returned).
-fn try_map<I: ParallelIterator, T: Send>(
-    iter: I,
-    f: impl Sync + Send + Fn(I::Item) -> Option<T>,
-) -> (Vec<T>, bool) {
-    let ok = AtomicBool::new(true);
-    let res = iter
-        .filter_map(|v| {
-            if !ok.load(Ordering::Acquire) {
-                return None;
-            }
-            let res = f(v);
-            if res.is_none() {
-                ok.store(false, Ordering::Release);
-            }
-            res
-        })
-        .collect();
-    (res, ok.load(Ordering::Acquire))
-}
 
 #[derive(thiserror::Error, Debug, PartialEq, Eq)]
 pub(crate) enum Error {
@@ -211,12 +146,13 @@ impl Cache {
 
         // Verify the signatures in parallel.
         // Verification will stop at the first encountered error.
-        let (data, ok) = try_map(data_and_keys.into_values().par_bridge(), |(d, key)| {
-            if d.payload().verify(&key).is_ok() {
-                return Some(d);
-            }
-            return None;
-        });
+        let (data, ok) =
+            concurrency::rayon::try_map(data_and_keys.into_values().par_bridge(), |(d, key)| {
+                if d.payload().verify(&key).is_ok() {
+                    return Some(d);
+                }
+                return None;
+            });
         if !ok {
             return (data, Some(Error::InvalidSignature));
         }
@@ -232,7 +168,7 @@ impl Cache {
     ) -> (Vec<Arc<SignedAccountData>>, Option<Error>) {
         let this = self.clone();
         // Execute verification on the rayon threadpool.
-        let (data, err) = rayon_spawn(move || this.verify(data)).await;
+        let (data, err) = concurrency::rayon::run(move || this.verify(data)).await;
         // Insert the successfully verified data, even if an error has been encountered.
         let inserted =
             self.0.update(|inner| data.into_iter().filter_map(|d| inner.try_insert(d)).collect());
